#647 Move quotedatabase code to databasemanager.ts move questiontypes and related types to questiontypes.ts add suggest quote button to all ciphers except cryptarithm
