+++ app/codebusters/cipheraffineencoder.ts
@@ -80,6 +80,7 @@ export class CipherAffineEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
