+++ app/codebusters/cipheraristocratencoder.ts
@@ -81,6 +81,7 @@ export class CipherAristocratEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
