+++ app/codebusters/ciphercheckerboardencoder.ts
@@ -112,6 +112,7 @@ export class CipherCheckerboardEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
