+++ app/codebusters/cipherdancingmenencoder.ts
@@ -34,6 +34,7 @@ export class CipherDancingMenEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
