+++ app/codebusters/cipherfractionatedmorseencoder.ts
@@ -101,6 +101,7 @@ export class CipherFractionatedMorseEncoder extends CipherMorseEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
