+++ app/codebusters/cipherhillencoder.ts
@@ -56,6 +56,7 @@ export class CipherHillEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
