+++ app/codebusters/cipherhomophonicencoder.ts
@@ -148,6 +148,7 @@ export class CipherHomophonicEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
