+++ app/codebusters/cipherknightstemplar.ts
@@ -30,6 +30,7 @@ export class CipherKnightsTemplarEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
