+++ app/codebusters/ciphermorseencoder.ts
@@ -18,6 +18,7 @@ export class CipherMorseEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
