+++ app/codebusters/ciphernihilistsubstitutionencoder.ts
@@ -121,6 +121,7 @@ export class CipherNihilistSubstitutionEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
