+++ app/codebusters/cipherpigpenencoder.ts
@@ -38,6 +38,7 @@ export class CipherPigPenEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
