+++ app/codebusters/cipherrailfenceencoder.ts
@@ -457,6 +457,7 @@ export class CipherRailFenceEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
