+++ app/codebusters/cipherstandardgalacticalphabetencoder.ts
@@ -32,6 +32,7 @@ export class CipherStandardGalacticAlphabetEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
