+++ app/codebusters/ciphertableencoder.ts
@@ -62,6 +62,7 @@ export class CipherTableEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
