+++ app/codebusters/ciphertapcodeencoder.ts
@@ -71,6 +71,7 @@ export class CipherTapCodeEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
