+++ app/codebusters/ciphertest.ts
@@ -26,35 +26,14 @@ import { pushLinkedTest, withCloudSyncSuppressed } from './cloudtestsync';
 import { waitForCloudAuth } from './cloudauth';
 import CipherScoreSheetGenerator from './cipherscoresheetgenerator';
 import { generateTestLatex } from './ciphertestlatexer';
-import { DifficultyType, TemplateTestTypes, TestDifficultyType, ITempleteInfo } from './ciphertestbuild';
+import { DatabaseManager } from './databasemanager'
 
-const DATABASE_VERSION = 4
+
+
 
 const OVERLEAF_TEMPLATE_URL = 'https://www.overleaf.com/read/yxjtjdcgykgp#aa9620';
 
-export interface QuestionType {
-    cipherType: ICipherType; // Type of cipher
-    group: number; // Which group the question belongs to
-    weight: number; // Likelihood that we want to use this on a test
-    timed?: boolean;  // This can be used as a timed question
-    lang?: string;// Optional language string
-    title: string; // Title to show user what type of question is being generated
-    testtype?: ITestType[];  // Tests that it is allowed to be on
-    shift?: number[]; // Range of shift for a caesar cipher
-    operation?: IOperationType; // Type of cipher opation (decode/crypt, etc)
-    encodeType?: IEncodeType;  // Type of encoding (random, K1, K2, K3) for an aristocrat
-    difficulty?: DifficultyType; // General difficulty for the question
-    keyword?: string; // Keyword to set hill length
-    guidance: string;
-    default?: string;
-    len?: number[]
-    chi2?: number[]
-    unique?: number[]
-    homonyms?: number[]
-    msg?: string;
-    usehint?: boolean;
-    misspelled?: boolean;
-}
+
 
 export interface SortableEntry {
     weight: number;
@@ -68,9 +47,7 @@ export interface buttonInfo {
     disabled?: boolean;
 }
 
-export interface UsedIdMap {
-    [index: number]: boolean;
-}
+
 
 export type modelID = string;
 export type IRealtimeObject = 'sourcemodel' | 'testmodel' | 'answertemplate' | 'answeraudit'; //| 'answermodel'
@@ -206,33 +183,14 @@ export interface IAnswerAudit {
 export type ITestDisp = 'testedit' | 'testprint' | 'ressheet' | 'testans' | 'testanstiny' | 'testsols' | 'testint';
 export type ITestManage = 'local' | 'published';
 
-export interface QuoteUpdates {
-    [id: number]: QuoteRecord
-}
 
-export interface DBTable {
-    Table: IDBObjectStore
-    Transaction: IDBTransaction
-    // LengthIdx?: IDBIndex
-    // Chi2Idx?: IDBIndex
-    // GradeIdx?: IDBIndex
-    // UniqueIdx?: IDBIndex
-}
 
-export interface QueryParms {
-    len?: number[]
-    chi2?: number[]
-    grade?: number[]
-    keywords?: string[]
-    unique?: number[]
-    homonyms?: number[]
-    testUsage?: boolean
-    start?: number
-    limit?: number
-}
+
+
+
 
 export type KeyRangeMap = Record<string, IDBKeyRange>;
-export const allButARegional = [ITestType.None, ITestType.bregional, ITestType.bstate, ITestType.cregional, ITestType.cstate]
+
 
 /**
  * Base support for all the test generation handlers
@@ -241,437 +199,7 @@ export class CipherTest extends CipherEncoder {
 
     public requiredSpecialBonusCipher: ICipherType = ICipherType.Homophonic
 
-    public questionChoices: QuestionType[] = [
-        // ----------------------------------------------------------------------------------------------
-        //    GROUP 1 - Aristocrats
-        // ----------------------------------------------------------------------------------------------
-        {
-            title: 'Easy Aristocrat with a Hint',
-            guidance: 'Easy Quote [83-100 non-blank characters, χ²<20] with Hint',
-            len: [83, 100], chi2: [-Infinity, 20], unique: [19, Infinity], usehint: true,
-            group: 1, weight: 0.75, cipherType: ICipherType.Aristocrat,
-            operation: 'decode', encodeType: 'random', difficulty: 'easy',
-        },
-        {
-            title: 'Easy Aristocrat without a Hint',
-            guidance: 'Easy Quote [83-100 non-blank characters, χ²<20]',
-            len: [83, 100], chi2: [-Infinity, 20], unique: [19, Infinity],
-            group: 1, weight: 0.75, timed: true, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'random', difficulty: 'easy',
-        },
-        {
-            title: 'Medium Aristocrat with a Hint',
-            guidance: 'Medium Quote [83-100 non-blank characters, 20<χ²<25] with Hint',
-            len: [83, 100], chi2: [20, 25], unique: [19, Infinity], usehint: true,
-            group: 1, weight: 0.75, cipherType: ICipherType.Aristocrat,
-            operation: 'decode', encodeType: 'random', difficulty: 'medium',
-        },
-        {
-            title: 'Medium Aristocrat without a Hint',
-            guidance: 'Medium Quote [83-100 non-blank characters, 20<χ²<25]',
-            len: [83, 100], chi2: [20, 25], unique: [19, Infinity],
-            group: 1, weight: 0.75, timed: true, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'random', difficulty: 'medium',
-        },
-        {
-            title: 'Hard K1 Aristocrat with a Hint',
-            guidance: 'Hard Quote [83-100 non-blank characters, χ²>25] with Hint',
-            len: [83, 100], chi2: [25, Infinity], unique: [19, Infinity], usehint: true,
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k1', difficulty: 'hard',
-        },
-        {
-            title: 'Hard K1 Aristocrat without a Hint',
-            guidance: 'Hard Quote [83-100 non-blank characters, χ²>25]',
-            len: [83, 100], chi2: [25, Infinity], unique: [19, Infinity],
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k1', difficulty: 'hard',
-        },
-        {
-            title: 'Hard K2 Aristocrat with a Hint',
-            guidance: 'Hard Quote [83-100 non-blank characters, χ²>25] with Hint',
-            len: [83, 100], chi2: [25, Infinity], unique: [19, Infinity], usehint: true,
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k2', difficulty: 'hard',
-        },
-        {
-            title: 'Hard K2 Aristocrat without a Hint',
-            guidance: 'Hard Quote [83-100 non-blank characters, χ²>25]',
-            len: [83, 100], chi2: [25, Infinity], unique: [19, Infinity],
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k2', difficulty: 'hard',
-        },
-        {
-            title: 'Misspelled K1 Aristocrat with a Hint',
-            guidance: 'Misspelled Quote [83-100 non-blank characters, χ²<25] with Hint',
-            len: [83, 100], chi2: [-Infinity, 25], unique: [19, Infinity], homonyms: [6, Infinity], usehint: true,
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat, misspelled: true,
-            testtype: [ITestType.None], operation: 'decode', encodeType: 'k1',
-        },
-        {
-            title: 'Misspelled K1 Aristocrat without a Hint',
-            guidance: 'Misspelled Quote [83-100 non-blank characters, χ²<25]',
-            len: [83, 100], chi2: [-Infinity, 25], unique: [19, Infinity], homonyms: [6, Infinity],
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat, misspelled: true,
-            testtype: [ITestType.None], operation: 'decode', encodeType: 'k1',
-        },
-        {
-            title: 'Misspelled K2 Aristocrat with a Hint',
-            guidance: 'Misspelled Quote [83-100 non-blank characters, χ²<25] with Hint',
-            len: [83, 100], chi2: [-Infinity, 25], unique: [19, Infinity], homonyms: [6, Infinity], usehint: true,
-            testtype: [ITestType.None], operation: 'decode', encodeType: 'k2',
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat, misspelled: true,
-        },
-        {
-            title: 'Misspelled K2 Aristocrat without a Hint',
-            guidance: 'Misspelled Quote [83-100 non-blank characters, χ²<25]',
-            len: [83, 100], chi2: [-Infinity, 25], unique: [19, Infinity], homonyms: [6, Infinity],
-            group: 1, weight: 0.5, cipherType: ICipherType.Aristocrat, misspelled: true,
-            testtype: [ITestType.None], operation: 'decode', encodeType: 'k2',
-        },
-        {
-            title: 'Keyword/Key Phrase K1 Aristocrat',
-            guidance: 'Medium Quote [83-100 non-blank characters, 20<χ²<25]',
-            len: [83, 100], chi2: [20, 25], unique: [19, Infinity],
-            testtype: allButARegional, operation: 'keyword', encodeType: 'k1', difficulty: 'medium',
-            group: 1, weight: 0.75, cipherType: ICipherType.Aristocrat,
-        },
-        {
-            title: 'Keyword/Key Phrase K2 Aristocrat',
-            guidance: 'Easy Quote [83-100 non-blank characters, χ²<20]',
-            len: [83, 100], chi2: [-Infinity, 20], unique: [19, Infinity],
-            group: 1, weight: 0.75, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'keyword', encodeType: 'k2', difficulty: 'easy',
-        },
-        {
-            title: 'Keyword/Key Phrase K3 Aristocrat',
-            guidance: 'Easy Quote [83-100 non-blank characters, χ²<20]',
-            len: [83, 100], chi2: [-Infinity, 20], unique: [19, Infinity],
-            group: 1, weight: 0.75, cipherType: ICipherType.Aristocrat,
-            testtype: allButARegional, operation: 'keyword', encodeType: 'k3', difficulty: 'easy',
-        },
-        {
-            title: 'Easy K1 Patristocrat',
-            guidance: 'Easy Quote [95-110 non-blank characters, χ²<20] with Hint',
-            len: [95, 110], chi2: [-Infinity, 20], unique: [19, Infinity], usehint: true,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k1', difficulty: 'easy',
-            group: 1, weight: 0.25, cipherType: ICipherType.Patristocrat,
-        },
-        {
-            title: 'Easy K2 Patristocrat',
-            guidance: 'Easy Quote [95-110 non-blank characters, χ²<20] with Hint',
-            len: [95, 110], chi2: [-Infinity, 20], unique: [19, Infinity], usehint: true,
-            group: 1, weight: 0.25, cipherType: ICipherType.Patristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k2', difficulty: 'easy',
-        },
-        {
-            title: 'Medium K1 Patristocrat',
-            guidance: 'Medium Quote [95-110 non-blank characters, 20<χ²<25] with Hint',
-            len: [95, 110], chi2: [20, 30], unique: [19, Infinity], usehint: true,
-            group: 1, weight: 0.25, cipherType: ICipherType.Patristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k1', difficulty: 'medium',
-        },
-        {
-            title: 'Medium K2 Patristocrat',
-            guidance: 'Medium Quote [95-110 non-blank characters, 20<χ²<25] with Hint',
-            len: [95, 110], chi2: [20, 30], unique: [19, Infinity], usehint: true,
-            group: 1, weight: 0.25, cipherType: ICipherType.Patristocrat,
-            testtype: allButARegional, operation: 'decode', encodeType: 'k2', difficulty: 'medium',
-        },
-        // ----------------------------------------------------------------------------------------------
-        //    GROUP 2 - Xenocrypts
-        // ----------------------------------------------------------------------------------------------
-        {
-            title: 'Medium K1 Spanish Xenocrypt',
-            guidance: 'Medium Spanish Quote [83-100 non-blank characters, 20<χ²<25]',
-            len: [83, 100], chi2: [20, 25], unique: [19, Infinity],
-            group: 2, weight: 0.5, cipherType: ICipherType.Aristocrat, lang: 'es',
-            encodeType: 'k1', difficulty: 'medium',
-        },
-        {
-            title: 'Medium K2 Spanish Xenocrypt',
-            guidance: 'Medium Spanish Quote [83-100 non-blank characters, 20<χ²<25]',
-            len: [83, 100], chi2: [20, 25], unique: [19, Infinity],
-            group: 2, weight: 0.5, cipherType: ICipherType.Aristocrat, lang: 'es',
-            encodeType: 'k2', difficulty: 'medium',
-        },
-        // ----------------------------------------------------------------------------------------------
-        //    GROUP 3 - Other Cipher types
-        // ----------------------------------------------------------------------------------------------
-        {
-            title: "Affine Decode",
-            guidance: '[25-30 non-blank characters, 13 or more unique letters]',
-            len: [25, 30], unique: [13, Infinity],
-            group: 3, weight: 0.5, cipherType: ICipherType.Affine,
-            operation: 'decode'
-        },
-        {
-            title: "Affine Cryptanalysis",
-            guidance: '[25-30 non-blank characters, 13 or more unique letters]',
-            len: [25, 30], unique: [13, Infinity],
-            group: 3, weight: 0.5, cipherType: ICipherType.Affine,
-            operation: 'crypt'
-        },
-        {
-            title: "Easy Caesar",
-            guidance: '[20-45 Characters.  Shift value +/- 3]',
-            len: [20, 45],
-            group: 3, weight: 0.5, cipherType: ICipherType.Caesar,
-            shift: [-3, 3], testtype: [ITestType.aregional], operation: 'decode'
-        },
-        {
-            title: "Caesar",
-            guidance: '[80-90 Characters. No single letter words]',
-            len: [80, 90],
-            group: 3, weight: 0.5, cipherType: ICipherType.Caesar,
-            testtype: allButARegional, operation: 'decode'
-        },
-        {
-            title: "Dancing Men",
-            guidance: '[20-30 Characters]',
-            len: [20, 30],
-            group: 3, weight: 0.5, cipherType: ICipherType.DancingMen,
-        },
-        {
-            title: "Standard Galactic Alphabet",
-            guidance: '[20-30 Characters]',
-            len: [20, 30],
-            group: 3, weight: 0.5, cipherType: ICipherType.StandardGalacticAlphabet,
-        },
-        {
-            title: 'Atbash',
-            guidance: '[45-80 Characters]',
-            len: [45, 80],
-            group: 3, weight: 0.5, cipherType: ICipherType.Atbash,
-            operation: 'decode'
-        },
-        {
-            title: "Complete Columnar",
-            guidance: '[45-90 Characters]',
-            len: [45, 90],
-            group: 3, weight: 0.5, cipherType: ICipherType.CompleteColumnar,
-        },
-        {
-            title: "Complete Columnar",
-            guidance: '[45-90 Characters]',
-            len: [45, 90],
-            group: 3, weight: 0.5, cipherType: ICipherType.CompleteColumnar,
-        },
-        {
-            title: "Cryptarithm",
-            guidance: '[Addition formula with a carry to make 1 or 2 digits obvious]',
-            default: 'EFFORT+PROJECT=ATTEMPT',
-            msg: "Use the Cryptarithm generator for formulas",
-            group: 3, weight: 0.5, cipherType: ICipherType.Cryptarithm,
-        },
-        {
-            title: "Cryptarithm",
-            guidance: '[Addition formula with a carry to make 1 or 2 digits obvious]',
-            default: 'EFFORT+PROJECT=ATTEMPT',
-            msg: "Use the Cryptarithm generator for formulas",
-            group: 3, weight: 0.5, cipherType: ICipherType.Cryptarithm,
-        },
-        {
-            title: "Hill 2x2",
-            guidance: '[19-23 (Odd) characters]',
-            len: [19, 23],
-            group: 3, weight: 0.5, operation: 'decode', cipherType: ICipherType.Hill,
-            keyword: "TEST"
-        },
-        {
-            title: "Hill 3x3",
-            guidance: '[20-28 (not multiple of 3) characters]',
-            len: [20, 28],
-            group: 3, weight: 0.5, operation: 'decode', cipherType: ICipherType.Hill,
-            keyword: "TEMPORARY"
-        },
-        {
-            title: "Porta Decode",
-            guidance: '[55-62 characters]',
-            len: [55, 62],
-            group: 3, weight: 0.5, cipherType: ICipherType.Porta,
-            operation: 'decode'
-        },
-        {
-            title: "Porta Cryptanalysis",
-            guidance: '[55-62 characters]',
-            len: [55, 62],
-            group: 3, weight: 0.5, cipherType: ICipherType.Porta,
-            operation: 'crypt'
-        },
-        {
-            title: "Homophonic Decode",
-            guidance: '[55-62 characters]',
-            len: [55, 62],
-            group: 3, weight: 0.5, cipherType: ICipherType.Homophonic,
-            operation: 'decode'
-        },
-        {
-            title: "Homophonic Cryptanalysis",
-            guidance: '[55-62 characters]',
-            len: [55, 62],
-            group: 3, weight: 0.5, cipherType: ICipherType.Homophonic,
-            operation: 'crypt'
-        },
-        {
-            title: "Checkerboard Decode",
-            guidance: '[55-75 characters]',
-            len: [55, 75], chi2: [20, Infinity],
-            group: 3, weight: 0.65, cipherType: ICipherType.Checkerboard,
-            operation: 'decode'
-        },
-        {
-            title: "Checkerboard Decode",
-            guidance: '[55-75 characters]',
-            len: [55, 75], chi2: [20, Infinity],
-            group: 3, weight: 0.5, cipherType: ICipherType.Checkerboard,
-            operation: 'decode'
-        },
-        {
-            title: "Checkerboard Cryptanalysis",
-            guidance: '[55-75 characters]',
-            len: [55, 75], chi2: [20, Infinity],
-            group: 3, weight: 0.65, cipherType: ICipherType.Checkerboard,
-            operation: 'crypt'
-        },
-        {
-            title: "Nihilist Decode",
-            guidance: '[40-55 characters]',
-            len: [40, 55],
-            group: 3, weight: 0.5, cipherType: ICipherType.NihilistSubstitution,
-            operation: 'decode'
-        },
-        {
-            title: "Nihilist Decode",
-            guidance: '[40-55 characters]',
-            len: [40, 55],
-            group: 3, weight: 0.5, cipherType: ICipherType.NihilistSubstitution,
-            operation: 'decode'
-        },
-        {
-            title: "Nihilist Cryptanalysis",
-            guidance: '[45-60 characters]',
-            len: [45, 60],
-            group: 3, weight: 0.5, cipherType: ICipherType.NihilistSubstitution,
-            operation: 'crypt'
-        },
-        {
-            title: "Vigenère Decode",
-            guidance: '[50-60 characters]',
-            len: [50, 60],
-            group: 3, weight: 0.5, cipherType: ICipherType.Vigenere,
-            operation: 'decode'
-        },
-        {
-            title: "Vigenère Cryptanalysis",
-            guidance: '[50-60 characters]',
-            len: [50, 60],
-            group: 3, weight: 0.5, cipherType: ICipherType.Vigenere,
-            operation: 'crypt'
-        },
-        {
-            title: 'Running Key',
-            guidance: '[50-60 characters]',
-            len: [50, 60],
-            group: 3, weight: 0.5, cipherType: ICipherType.RunningKey,
-        },
-        {
-            title: "Baconian Letter for Letter",
-            guidance: '[35-50 characters]',
-            len: [35, 50],
-            group: 3, weight: 0.5, cipherType: ICipherType.Baconian,
-            operation: 'let4let'
-        },
-        {
-            title: "Baconian Sequence",
-            guidance: '[35-50 characters]',
-            len: [35, 50],
-            group: 3, weight: 0.5, cipherType: ICipherType.Baconian,
-            operation: 'sequence',
-        },
-        {
-            title: "Baconian Words",
-            guidance: '[30-40 characters]',
-            len: [30, 40],
-            group: 3, weight: 0.5, cipherType: ICipherType.Baconian,
-            operation: 'words'
-        },
-        {
-            title: "RSA",
-            guidance: '[2 digit primes]',
-            group: 3, weight: 0.5, cipherType: ICipherType.RSA,
-        },
-        {
-            title: "Pig Pen",
-            guidance: '[40-50 characters]',
-            len: [40, 50],
-            group: 3, weight: 0.5, cipherType: ICipherType.PigPen,
-        },
-        {
-            title: "Knights Templar",
-            guidance: '[40-50 characters]',
-            len: [40, 50],
-            group: 3, weight: 0.5, cipherType: ICipherType.KnightsTemplar,
-        },
-        {
-            title: 'TapCode',
-            guidance: '[18-30 characters]',
-            len: [18, 30],
-            group: 3, weight: 0.5, cipherType: ICipherType.TapCode,
-        },
-        {
-            title: "Morbit Decode",
-            guidance: '[35-42 characters]',
-            len: [35, 42],
-            group: 3, weight: 0.5, cipherType: ICipherType.Morbit,
-            operation: 'decode',
-        },
-        {
-            title: "Morbit Cryptanalysis",
-            guidance: '[35-42 characters]',
-            len: [35, 42],
-            group: 3, weight: 0.5, cipherType: ICipherType.Morbit,
-            operation: 'crypt',
-        },
-        {
-            group: 3, weight: 0.5, cipherType: ICipherType.Pollux,
-            title: "Pollux Decode",
-            len: [35, 42],
-            guidance: '[35-42 characters]',
-            operation: 'decode'
-        },
-        {
-            group: 3, weight: 0.5, cipherType: ICipherType.Pollux,
-            title: "Pollux Cryptanalysis",
-            len: [35, 42],
-            guidance: '[35-42 characters]',
-            operation: 'crypt'
-        },
-        {
-            title: "Fractionated Morse",
-            len: [38, 52],
-            guidance: '[38-52 characters]',
-            group: 3, weight: 0.5, cipherType: ICipherType.FractionatedMorse,
-        },
-        {
-            title: "Railfence Variable Rails",
-            guidance: '[55-62 characters]',
-            len: [55, 62],
-            group: 3, weight: 0.5, cipherType: ICipherType.Railfence,
-        },
-        {
-            title: "Railfence Fixed Rails",
-            guidance: '[55-62 characters]',
-            len: [55, 62],
-            group: 3, weight: 0.5, cipherType: ICipherType.Railfence,
-        },
-        {
-            title: "Railfence Variable Rails and Offset",
-            guidance: '[55-62 characters]',
-            len: [55, 62],
-            group: 3, weight: 0.5, cipherType: ICipherType.Railfence,
-        },
-    ];
+
 
     public readonly cipherSubTypes = [
         'Aristocrat',
@@ -821,7 +349,6 @@ export class CipherTest extends CipherEncoder {
      * Any running keys used for the test
      */
     public runningKeys: IRunningKey[];
-    public LocalDB: IDBOpenDBRequest = undefined
     /**
      * Restore the state from either a saved file or a previous undo record
      * @param data Saved state to restore
@@ -1479,47 +1006,8 @@ export class CipherTest extends CipherEncoder {
         }
         return "english"
     }
-    public CreateTableIfNeeded(db: IDBDatabase, lang: string): void {
-        if (!db.objectStoreNames.contains(lang)) {
-            const Table = db.createObjectStore(lang, { keyPath: "id", autoIncrement: true });
-            Table.createIndex('minquote', 'minquote', { unique: true });
-            Table.createIndex('len', 'len')
-            Table.createIndex('chi2', 'chi2')
-            Table.createIndex('grade', 'grade')
-            Table.createIndex('unique', 'unique')
-        }
-    }
-    /**
-     * 
-     * @param lang 
-     * @returns 
-     */
-    public openDatabase(lang: string, mode: IDBTransactionMode = "readonly"): Promise<DBTable> {
-        return new Promise<DBTable>((resolve, reject) => {
-            this.LocalDB = window.indexedDB.open("cipher_quotes", DATABASE_VERSION);
-            this.LocalDB.onerror = (ev) => {
-                reject(`Unable to open database: ${(ev.target as IDBOpenDBRequest).error}`)
-            }
-            this.LocalDB.onsuccess = (ev) => {
-                const db = (ev.target as IDBOpenDBRequest).result;
-                const transaction = db.transaction(lang, mode)
-                if (transaction === undefined) {
-                    reject(`Unable to open database: ${(ev.target as IDBOpenDBRequest).error}`)
-                } else {
-                    resolve({ Table: transaction.objectStore(lang), Transaction: transaction })
-                }
-            }
-            this.LocalDB.onupgradeneeded = (ev) => {
-                console.log('Database needs to be upgraded')
 
-                const db = (ev.target as IDBOpenDBRequest).result;
-                db.onerror = (evt) => { console.log(`Database error: ${(evt.target as IDBOpenDBRequest).error}`) }
-                // Create an objectStore for this database
-                this.CreateTableIfNeeded(db, "english")
-                this.CreateTableIfNeeded(db, "spanish")
-            }
-        })
-    }
+
     /**
      * 
      * @param step Step number to display
@@ -1530,351 +1018,8 @@ export class CipherTest extends CipherEncoder {
         const title = $('<h3>').text(step)
         return makeCallout($("<div/>").append(title).append(body as HTMLElement), 'secondary')
     }
-    /**
-     * Make a range that can be passed to IndexDB openCursor routines
-     * From https://developer.mozilla.org/en-US/docs/Web/API/IDBKeyRange
-     *       All keys ≥ x         IDBKeyRange.lowerBound(x)
-     *       All keys > x         IDBKeyRange.lowerBound(x, true)
-     *       All keys ≤ y         IDBKeyRange.upperBound(y)
-     *       All keys < y         IDBKeyRange.upperBound(y, true)
-     *       All keys ≥ x && ≤ y  IDBKeyRange.bound(x, y)
-     *       All keys > x && < y  IDBKeyRange.bound(x, y, true, true)
-     *       All keys > x && ≤ y  IDBKeyRange.bound(x, y, true, false)
-     *       All keys ≥ x &&< y   IDBKeyRange.bound(x, y, false, true)
-     *       The key = z          IDBKeyRange.only(z)
-     * @param rangeVals Pair of range numbers.  Undefined for either is unlimited
-     * @returns IDBKeyRange object that can be passed to openCursor
-     */
-    public makeRange(rangeVals: number[]): IDBKeyRange {
-        let lower = rangeVals[0]
-        let upper = rangeVals[1]
-        // Do we have a lower bound?
-        if (lower === -Infinity) {
-            // No Lower bound, how about an upper bound?
-            if (upper === Infinity) {
-                // No bounds, so we just get everything
-                return undefined
-            }
-            // Just an upper bound
-            return IDBKeyRange.upperBound(upper)
-        }
-        // We have a lower bound, How about an upper bound too?
-        if (upper === Infinity) {
-            // No upper bound
-            return IDBKeyRange.lowerBound(lower)
-        }
-        // We have a special case where both values are the same
-        if (lower === upper) {
-            return IDBKeyRange.only(lower)
-        }
-        // Otherwise we have an inclusive range for the lower/upper bounds
-        return IDBKeyRange.bound(lower, upper)
-    }
-    /**
-     * Convert all undefined values in a range to the appropriate Infinity range
-     * @param range Range to fix
-     * @returns Updated range
-     */
-    public fixRange(range: number[]): number[] {
-        if (range === undefined) {
-            return [-Infinity, Infinity]
-        }
-        let lower = range[0] ?? -Infinity
-        let upper = range[1] ?? Infinity
-        if (lower <= upper) {
-            return [lower, upper]
-        }
-        return [upper, lower]
-    }
-    /**
-     * Normalize all the parameters, setting ranges for any undefined values
-     * @param parmsReq Parameters structure of limits on query
-     * @returns Cleaned up set of parameters
-     */
-    public cleanParms(parmsReq: QueryParms): QueryParms {
-        const result: QueryParms = {}
-        result.start = parmsReq.start ?? 0;
-        result.limit = parmsReq.limit ?? 50;
-        result.chi2 = this.fixRange(parmsReq.chi2)
-        result.len = this.fixRange(parmsReq.len)
-        result.grade = this.fixRange(parmsReq.grade)
-        result.unique = this.fixRange(parmsReq.unique)
-        result.homonyms = this.fixRange(parmsReq.homonyms)
-        result.testUsage = parmsReq.testUsage
-        // Copy over any keywords
-        if (parmsReq.keywords !== undefined && parmsReq.keywords.length > 0) {
-            result.keywords = parmsReq.keywords
-        } else {
-            result.keywords = undefined
-        }
-        return result
-    }
-    /**
-     * Search all entries which match a given set of criteria and call a processing routine for each matched entry
-     * @param lang Language to search
-     * @param parmsReq Filters to apply the the search
-     * @param process Routine to process any matched records.  If this routine returns true, the search is ended early.
-     * @returns Success/failure boolean.
-     */
-    public async SearchEntriesWithRanges(lang: string, parmsReq: QueryParms, process: (rec: QuoteRecord) => boolean): Promise<boolean> {
-        const parms = this.cleanParms(parmsReq)
 
-        return new Promise<boolean>((resolve, reject) => {
-            this.openDatabase(lang).then((db) => {
-                try {
-                    const transaction = db.Transaction
-                    const store = transaction.objectStore(lang);
-                    // Figure out what type of cursor we will have for 
-                    let cursorRequest: IDBRequest<IDBCursorWithValue>
 
-                    // We only get to use one key for IndexDB, so let's
-                    // pick the ones which are the most likely to filter down
-                    let idxname = 'len'
-                    let rangeType = this.makeRange(parms.len)
-                    if (rangeType === undefined) {
-                        idxname = 'chi2'
-                        rangeType = this.makeRange(parms.chi2)
-                    }
-                    if (rangeType === undefined) {
-                        idxname = 'grade'
-                        rangeType = this.makeRange(parms.grade)
-                    }
-                    if (rangeType === undefined) {
-                        idxname = 'unique'
-                        const rangeType = this.makeRange(parms.unique)
-                    }
-                    // If one of those succeeded then we open the index on that field
-                    if (rangeType !== undefined) {
-                        const idx = store.index(idxname)
-                        cursorRequest = idx.openCursor(rangeType)
-                    } else {
-                        // Otherwise no filters, so just use the main cursor
-                        cursorRequest = store.openCursor()
-                    }
-                    const entries: QuoteRecord[] = [];
-                    let current = -1;
-
-                    cursorRequest.onsuccess = (event) => {
-                        const cursor = cursorRequest.result;
-                        if (cursor) {
-                            // See if this is a valid entry
-                            if (this.matchesRange(cursor.value, parms)) {
-                                // It matches, so see if we need to account for it
-                                current++
-                                // if (current % 1000 === 0) {
-                                //     console.log(`Checked ${current} entries, found ${entries.length} matches so far...`)
-                                // }
-                                // Are we past the ones we should skip? 
-                                if (current >= parms.start) {
-                                    // Yes, so remember it
-                                    if (process(cursor.value))
-                                        resolve(true)
-                                }
-                            }
-                            cursor.continue();
-                        } else {
-                            // No more records, resolve the Promise
-                            resolve(true);
-                        }
-                    };
-
-                    cursorRequest.onerror = (event) => {
-                        reject(`Error reading records: ${(event.target as IDBRequest).error}`);
-                    };
-                } catch (e) {
-                    return resolve(false);
-                }
-            }).catch((_e) => { return resolve(false) });
-        })
-    }
-    /**
-     * Get all the entries which match a given range.
-     * @param lang Language database to search
-     * @param parmsReq Ranges of entries to filter against
-     * @param limit Maximum number of entries to return
-     * @returns Promise to array of QuoteRecords
-     */
-    public async getEntriesWithRanges(lang: string, parmsReq: QueryParms, limit: number = 25): Promise<QuoteRecord[]> {
-        const parms = this.cleanParms(parmsReq)
-
-        return new Promise<QuoteRecord[]>((resolve, reject) => {
-
-            const entries: QuoteRecord[] = [];
-            this.SearchEntriesWithRanges(lang, parmsReq, (entry: QuoteRecord): boolean => {
-                entries.push(entry);
-                return (entries.length > limit);
-            }).then((res: boolean) => {
-                resolve(entries);
-            }).catch((reason) => { reject(reason) });
-        })
-    }
-    /**
-     * Get a random set of entries which match a given criteria.
-     * @param lang Language database to search
-     * @param parmsReq Ranges of entries to filter against
-     * @param limit Maximum number of entries to return
-     * @returns Promise to array of QuoteRecords
-     */
-    public async getRandomEntriesWithRanges(lang: string, parmsReq: QueryParms, usedmap: UsedIdMap, limit: number = 3): Promise<QuoteRecord[]> {
-        const parms = this.cleanParms(parmsReq)
-
-        return new Promise<QuoteRecord[]>((resolve, reject) => {
-
-            const entries: QuoteRecord[] = [];
-            const weights: number[] = [];
-            let maxweight = -1;
-            this.SearchEntriesWithRanges(lang, parmsReq, (entry: QuoteRecord): boolean => {
-                if ((entry.testUsage === undefined || entry.testUsage === "") && usedmap[entry.id] !== true) {
-                    const weight = Math.random()
-                    if (entries.length < limit) {
-                        entries.push(entry);
-                        weights.push(weight);
-                        if (weight > maxweight) {
-                            maxweight = weight
-                        }
-                        return (false)
-                    }
-                    // We have more than the limit, so figure out which one to throw out (if any)
-                    if (weight < maxweight) {
-                        let toreplace = maxweight
-                        maxweight = weight;
-                        for (let i in entries) {
-                            if (weights[i] === toreplace) {
-                                // this is the one to replace
-                                weights[i] = weight
-                                entries[i] = entry
-                                // Since we replaced an entry, don't search for another to replace (on the off chance we get two random numbers of the same value)
-                                toreplace = -1;
-                            } else {
-                                if (weights[i] > maxweight) {
-                                    maxweight = weights[i]
-                                }
-                            }
-                        }
-                    }
-                }
-                return (false);
-            }).then((res: boolean) => {
-                resolve(entries);
-            }).catch((reason) => { reject(reason) });
-        })
-    }
-    /**
-     * Update a single record in the database.  Note that due to the oddities of the
-     * indexeddb database, we have to find the record, make a copy of it, modify the copy, delete the
-     * original record, remove the id from the new record and then add it back to the database.
-     * @param objectStore Database to update the record
-     * @param id Id of record to update
-     * @param toUpdate New data to apply to the record
-     * @returns 
-     */
-    public updateDBRecord(objectStore: IDBObjectStore, id: number, toUpdate: QuoteRecord): Promise<boolean> {
-        return new Promise<boolean>((resolve, reject) => {
-            // let cursorRequest: IDBRequest<IDBCursorWithValue>
-            const idKey = IDBKeyRange.only(id)
-            const cursorRequest = objectStore.openCursor(idKey)
-
-            cursorRequest.onsuccess = (event) => {
-                const cursor = cursorRequest.result;
-                if (cursor) {
-                    console.log(`Found Record at ${cursor.value.id}`)
-                    const updateData = cursor.value;
-                    for (let key in toUpdate) {
-                        updateData[key] = toUpdate[key]
-                    }
-                    delete updateData['id']
-
-                    const request = cursor.delete();
-                    request.onsuccess = () => {
-                        // Now we need to add it
-                        const request = objectStore.add(updateData)
-                        request.onsuccess = (event) => {
-                            resolve(true)
-                        }
-                        request.onerror = (event) => {
-                            reject(event);
-                        }
-                    };
-                    request.onerror = (ev) => {
-                        console.log(`Cursor delete failure`)
-                        console.log(ev)
-                        reject(ev)
-                    }
-                } else {
-                    // No more records, resolve the Promise
-                    resolve(true);
-                }
-            };
-
-            cursorRequest.onerror = (event) => {
-                reject(`Error reading records: ${(event.target as IDBRequest).error}`);
-            };
-        })
-    }
-    /**
-     * Update a series of entries in the database
-     * Primarily used to mark records as being used on a test
-     * @param lang Language database to update
-     * @param updatereq List of entries to be updated
-     * @returns 
-     */
-    public updateDBRecords(lang: string, updatereq: QuoteUpdates): Promise<boolean> {
-
-        return new Promise<boolean>((resolve, reject) => {
-            if (Object.keys(updatereq).length <= 0) {
-                resolve(true);
-                return;
-            }
-            this.openDatabase(lang, 'readwrite').then(async (db) => {
-                const objectStore = db.Table;
-                for (const id in updatereq) {
-                    await this.updateDBRecord(objectStore, Number(id), updatereq[id])
-                }
-                resolve(true);
-                return;
-            })
-        })
-
-    }
-    /**
-     * Determine if a record matches the filter requirements
-     * @param entry Entry to check
-     * @param parms Limits to check entry against
-     * @returns True if the record is a valid match
-     */
-    public matchesRange(entry: QuoteRecord, parms: QueryParms): boolean {
-        let result = (entry.chi2 >= parms.chi2[0] && entry.chi2 <= parms.chi2[1] &&
-            entry.len >= parms.len[0] && entry.len <= parms.len[1] &&
-            entry.grade >= parms.grade[0] && entry.grade <= parms.grade[1] &&
-            entry.unique >= parms.unique[0] && entry.unique <= parms.unique[1]
-        )
-        if (parms.testUsage !== undefined) {
-            let used = entry.testUsage !== undefined && entry.testUsage !== ""
-            if (parms.testUsage !== used) {
-                result = false
-            }
-        }
-        // If we have a keyword filter, see if any of the words appear in the quote or the notes.
-        if (result && (parms.keywords !== undefined)) {
-            let look = `${entry.quote} ${entry.notes}`.toLowerCase()
-            result = false
-            for (const keyword of parms.keywords) {
-                if (look.includes(keyword)) {
-                    result = true;
-                    break;
-                }
-            }
-        }
-        // If we are also looking for homonyms, 
-        if (result && (parms.homonyms !== undefined) && (parms.homonyms[0] > -Infinity)) {
-            const homonymcount = countHomonyms(entry.quote)
-            // If there were enough homonyms in the phrase we can keep it
-            if (homonymcount < parms.homonyms[0]) {
-                result = false;
-            }
-        }
-        return result
-    }
     /**
      * Generate a table showing all the questions with action buttons
      * @param filter Items not to be displayed (doesn't currently work)
