+++ app/codebusters/ciphervigenereencoder.ts
@@ -114,6 +114,7 @@ export class CipherVigenereEncoder extends CipherEncoder {
         this.undocmdButton,
         this.redocmdButton,
         this.questionButton,
+        this.quoteButton,
         this.pointsButton,
         this.guidanceButton,
     ];
