+++ app/common/styles.css
@@ -335,6 +335,7 @@ button.qmore {
 
 button.use,
 button.useq,
+button.usequote,
 button.ban {
     font-size: 0.8 rem;
     padding: 4px 6px;
